Edycja persona działa :)

+++ Drzewo_genealogiczne/Data/engine/C_db.cpp
@@ -66,7 +66,7 @@ void C_db::m_sort(bool(*F)(C_person_base* _left, C_person_base* _right)) {
 void C_db::m_add_person(bool b_SEX,std::string& s_first, std::string& s_last, C_date& d_brith, C_date& d_deadth) {
 	L_person.push_back(new C_person_null(ID_MAIN_++,b_SEX,s_first, s_last, d_brith, d_deadth));
 }
-void C_db::m_update_person(bool b_SEX,std::string& s_first, std::string& s_last, C_date& d_brith, C_date& d_deadth, int& i_variable) { //do szcegolnego przetestowania
+void C_db::m_update_person(bool b_SEX,std::string& s_first, std::string& s_last, C_date& d_brith, C_date& d_deadth, long long& i_variable) { //do szcegolnego przetestowania
 	std::list<C_person_base*>::iterator it = L_person.begin();
 	advance(it, i_variable); //przesuwa iterator listy o wartosc i_variable
 	//std::vector<C_relation> V_r;
